Fixup line of sight
The previous implementation looked at a bounding box between the player
and the monster. We want only to look for walls on the "straight" line
between the player and the monster. Of course, this is difficult with
the cells being squares, so interpolate creates a line as straight as
possible.